refactor: new file for route match request logic
